requirements.txt
openai
python-dotenv
pandas
pyarrow
langchain
langchain-openai
langchain-community
langchain_experimental
tabulate